make it better
see the linked issue/pr for more information on motivation
