+++ src/detector.ts
@@ -1,15 +1,5 @@
-import { detectors } from './detectors'
-import { sources } from './sources'
-import {
-  BotdError,
-  BotDetectionResult,
-  BotDetectorInterface,
-  BotKind,
-  Component,
-  ComponentDict,
-  DetectionDict,
-  State,
-} from './types'
+import { BotDetectionResult, BotDetectorInterface, ComponentDict, DetectionDict } from './types'
+import { collect, detect } from './api'
 
 /**
  * Class representing a bot detector.
@@ -30,16 +20,6 @@ export default class BotDetector implements BotDetectorInterface {
     return this.detections
   }
 
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-  protected getSources() {
-    return sources
-  }
-
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-  protected getDetectors() {
-    return detectors
-  }
-
   /**
    * @inheritdoc
    */
@@ -48,35 +28,9 @@ export default class BotDetector implements BotDetectorInterface {
       throw new Error("BotDetector.detect can't be called before BotDetector.collect")
     }
 
-    const components = this.components
-    const detectors = this.getDetectors()
-
-    const detections = {} as DetectionDict
-    let finalDetection: BotDetectionResult = {
-      bot: false,
-    }
-
-    for (const detectorName in detectors) {
-      const detector = detectors[detectorName as keyof typeof detectors]
-      const detectorRes = detector(components)
-
-      let detection: BotDetectionResult = { bot: false }
-
-      if (typeof detectorRes === 'string') {
-        detection = { bot: true, botKind: detectorRes }
-      } else if (detectorRes) {
-        detection = { bot: true, botKind: BotKind.Unknown }
-      }
-
-      detections[detectorName as keyof typeof detectors] = detection
-
-      if (detection.bot) {
-        finalDetection = detection
-      }
-    }
+    const [detections, finalDetection] = detect(this.components)
 
     this.detections = detections
-
     return finalDetection
   }
 
@@ -84,37 +38,7 @@ export default class BotDetector implements BotDetectorInterface {
    * @inheritdoc
    */
   public async collect(): Promise<ComponentDict> {
-    const sources = this.getSources()
-    const components = {} as ComponentDict
-
-    const sourcesKeys = Object.keys(sources) as (keyof typeof sources)[]
-
-    await Promise.all(
-      sourcesKeys.map(async (sourceKey) => {
-        const res = sources[sourceKey]
-
-        try {
-          components[sourceKey] = ({
-            value: await res(),
-            state: State.Success,
-          } as Component<any>) as any
-        } catch (error) {
-          if (error instanceof BotdError) {
-            components[sourceKey] = {
-              state: error.state,
-              error: `${error.name}: ${error.message}`,
-            }
-          } else {
-            components[sourceKey] = {
-              state: State.UnexpectedBehaviour,
-              error: error instanceof Error ? `${error.name}: ${error.message}` : String(error),
-            }
-          }
-        }
-      }),
-    )
-
-    this.components = components
+    this.components = await collect()
     return this.components
   }
 }
